Move features to top level folder & Improve split layouts code
