+++ src/features/code-editor/hooks/useEditor.ts
@@ -3,8 +3,7 @@ import { HighlightStyle, syntaxHighlighting } from "@codemirror/language";
 import { Compartment, EditorState, Extension } from "@codemirror/state";
 import { EditorView, ViewUpdate, keymap, lineNumbers } from "@codemirror/view";
 import { tags } from "@lezer/highlight";
-import { useEffect, useMemo, useRef, useState } from "react";
-import { Theme, useTheme } from "~/context/theme";
+import { useEffect, useRef, useState } from "react";
 import { graphene } from "~/core/graphene/tools/codeMirror";
 
 // Creates onChange extension for editor.
@@ -35,23 +34,6 @@ const codeThemeLight = HighlightStyle.define([
   },
 ]);
 
-// const codeThemeDark = HighlightStyle.define([
-//   {
-//     tag: tags.brace,
-//     color: "#CAD1D8",
-//   },
-//   {
-//     tag: tags.variableName,
-//     color: "#D2A8FF",
-//   },
-//   {
-//     tag: tags.number,
-//     color: "#D2A8FF",
-//   },
-// ]);
-
-const codeThemeDark = codeThemeLight;
-
 const editorThemeLight = EditorView.theme({
   "&": {
     backgroundColor: "#f9fafb",
@@ -63,41 +45,11 @@ const editorThemeLight = EditorView.theme({
   },
 });
 
-// const editorThemeDark = EditorView.theme(
-//   {
-//     "&": {
-//       backgroundColor: "#1E1F23",
-//       height: "100%",
-//     },
-//     ".cm-gutters": {
-//       backgroundColor: "#1E1F23",
-//     },
-//   },
-//   { dark: true }
-// );
-
-const editorThemeDark = editorThemeLight;
-
 export const useEditor = <T extends HTMLElement>(extensions: Extension[]) => {
   const ref = useRef<T>(null);
   const [view, setView] = useState<EditorView>();
 
-  const { theme } = useTheme();
   const editorThemeCompartment = useRef(new Compartment());
-
-  const themes: Record<Theme, Extension[]> = useMemo(
-    () => ({
-      light: [
-        editorThemeLight,
-        syntaxHighlighting(codeThemeLight, { fallback: true }),
-      ],
-      dark: [
-        editorThemeDark,
-        syntaxHighlighting(codeThemeDark, { fallback: true }),
-      ],
-    }),
-    []
-  );
 
   useEffect(() => {
     if (!ref.current) return;
@@ -109,7 +61,10 @@ export const useEditor = <T extends HTMLElement>(extensions: Extension[]) => {
           history(),
           keymap.of(defaultKeymap),
           lineNumbers(),
-          editorThemeCompartment.current.of([editorThemeLight]),
+          editorThemeCompartment.current.of([
+            editorThemeLight,
+            syntaxHighlighting(codeThemeLight, { fallback: true }),
+          ]),
           ...extensions,
         ],
       }),
@@ -125,15 +80,6 @@ export const useEditor = <T extends HTMLElement>(extensions: Extension[]) => {
 
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [ref]);
-
-  useEffect(() => {
-    // If the view is not initialized or valid theme is used do not change anything.
-    if (!view) return;
-
-    view.dispatch({
-      effects: editorThemeCompartment.current.reconfigure(themes[theme]),
-    });
-  }, [theme, themes, view]);
 
   return { view, ref };
 };
