+++ src/features/graph-view/components/Vertex.tsx
