+++ src/features/graph-view/hooks/useSvgControls.ts
