+++ src/features/graph-view/types/selectedVertex.ts
