+++ src/pages/editor/Editor.tsx
@@ -1,4 +1,3 @@
-import { useTheme } from "~/context/theme";
 import { VerticalSplit } from "~/shared/layout/VerticalSplit";
 import { Header } from "./components/Header";
 import { Sidebar } from "./components/Sidebar";
@@ -6,13 +5,11 @@ import { Visualizer } from "./components/Visualizer";
 import { EditorStoreProvider } from "./context/editor";
 
 export const Editor = () => {
-  const { themeClass } = useTheme();
-
   return (
     <EditorStoreProvider projectId={""} loadingFallback={<>Loading ...</>}>
-      <div className={`flex h-screen flex-col overflow-hidden ${themeClass}`}>
+      <div className={`flex h-screen flex-col`}>
         <Header name={""} onRename={undefined} />
-        <main className="h-[calc(100vh_-_48px)]">
+        <main className="min-h-0 flex-grow">
           <VerticalSplit left={<Sidebar />} right={<Visualizer />} />
         </main>
       </div>
