+++ src/pages/editor/components/Sidebar.tsx
@@ -1,6 +1,6 @@
 import { CommandLineIcon, PlayIcon } from "@heroicons/react/24/outline";
 import { Tab, VerticalTabs } from "~/shared/layout/VerticalTab";
-import { CodeEditor } from "../features/code-editor/CodeEditor";
+import { CodeEditor } from "../../../features/code-editor/CodeEditor";
 import { AlgorithmPicker } from "./AlgorithmPicker";
 
 const sidebarTabs: Tab[] = [
